Add requestjs as dependency for Dispatcher
This allows Dispatcher to use the correct "require" instance in case of
multiple-context environments.

Example: When using Aloha editor in a Chaplin app, Aloha is being
loaded in a different requirejs context. Chaplin thus uses require with
a wrong baseUrl setting and cannot find controllers.

Adding require as dependency for Dispatcher fixes that.