Allow regular expressions in filter definitions, don't override filters in controller subclasses anymore
@rendez and I have been discussing this morning about a few things:

First off, we agree with @mehcode that using RegExps for filter names is the right choice.

Looking back at Rails we considered the options `:only` and and `:except` for before filters. This flexibility is nice, but then that would require for example an excluding array, which then we have to decide whether it would be for all prototype chains or just the current one, etc.
We do still believe Regexps must be used carefully, as they may be prone to mistakes, but do offer all the flexibility to really scope the named actions one wants to catch with a before filter definition, in order to have extended controllers take advantage of it.

Secondly, if two controllers have the same name for a before filter, the second filter does not override the first one anymore, but both filters are executed. As shown by @paulmillr's `delegateEvents` code, this is easy to achieve. We've copied the getPrototypeChain function from 06eb37d for this.

We see the attached code state as minimal viable version of this feature that can be improved on a constant basis; this fits well into Chaplin's development cycle.