Docco annotated source code.
We should produce and publish annotated source code on the chaplinjs.org web site.  This is **application.coffee** and if it meets with your agreement I'll do the rest.  Also, files in the `/docs` directory should link to these newly produced pages via "The Code" hyperlink currently at the end of each.