Fix internal links
I am not entirely happy with the relative links as they are now.
Relative links were brittle already with the use of GitHub's section
links, where links would break with changing method signatures. This
situation has now become even worse, where unrelated changes, e.g. an
additional method being added to the docs, can break links.

This is because I use RedCarpet's feature of automatically creating
heading identifiers and it doesn't use the text of the heading to do so.
Maybe going for HTML with IDs here is actually the most reliable choice.