Improve method/property naming in docs
I decided to use "this.foo" instead of "@foo" when it appears
interspersed in text to make it work for both CoffeeScript and
JavaScript.

However, I also realized that in many cases it makes sense to use simply
"foo", since while it may be a property of the class of objects being
discussed, there is no reason to always present it as being accessed
from within the object.