Fix bug with double-freezing of Application
Application was frozen after `start()` called, this leads to exception on attempt to dispose the Application, at that time Application is already frozen so setting `disposed = true` raises TypeError in browser. Changing freezing to sealing instead.
The unit tests were updated to check if Application is correctly sealed on start.